+++ PROYECTO-01-DOMINGUEZ-JUAN.py
@@ -3,7 +3,7 @@ PROJECT 1
 This is a basic python based data science solution
 author: juan_dm93@hotmail.com
 """
-from utils.frontend import login
+from frontend import login
 
 
 def main():
