+++ frontend/locals.py
@@ -1,0 +1,15 @@
+# LOCALS
+SLEEPING = 0.4
+PRINT_SIZE = 10
+USERS = {
+    'admin': 'pass',
+}
+
+PROCESSES = ['globals', 'cats', 'date']
+EXIT_CMDS = ['return', 'logout', 'exit']
+
+MONTHS = [
+    'JAN', 'FEB', 'MAR', 'APR',
+    'MAY', 'JUN', 'JUL', 'AUG',
+    'SEP', 'OCT', 'NOV', 'DEC',
+]
